Nomina v1.0.0.54

+++ Altas.Core/Altas.cs
@@ -22,7 +22,7 @@ namespace Altas.Core
         public int jornada { get; set; }
         public DateTime fechaingreso { get; set; }
         public int diasproporcionales { get; set; }
-        public double sdi { get; set; }
+        public decimal sdi { get; set; }
         public string cp { get; set; }
         public DateTime fechanacimiento { get; set; }
         public string estado { get; set; }
