+++ Imss.Core/Imss.cs
@@ -10,6 +10,6 @@ namespace Imss.Core
     {
         public int id { get; set; }
         public string prestacion { get; set; }
-        public double porcentaje { get; set; }
+        public decimal porcentaje { get; set; }
     }
 }
