+++ Movimientos.Core/Movimientos.cs
@@ -12,7 +12,7 @@ namespace Movimientos.Core
         public int idtrabajador { get; set; }
         public int idempresa { get; set; }
         public int idconcepto { get; set; }
-        public double cantidad { get; set; }
+        public decimal cantidad { get; set; }
         public DateTime fechainicio { get; set; }
         public DateTime fechafin { get; set; }
     }
