+++ ProgramacionConcepto.Core/ProgramacionConcepto.cs
@@ -13,7 +13,7 @@ namespace ProgramacionConcepto.Core
         public int idempresa { get; set; }
         public int idconcepto { get; set; }
         public string concepto { get; set; }
-        public double cantidad { get; set; }
+        public decimal cantidad { get; set; }
         public DateTime fechafin { get; set; }
     }
 }
